Move ImageStoreSpy to shared scope
